Fix JaninoMetaTest
It breaks when adding in a new ValueMeta type because of how it's
selecting a random type id assuming a compact id space. Instead, select
a random type from the list of type name strings and then fetch the type
id from the factory. Same functional result preserving the test logic.